Change `getMostUsed`’s failure value to `null`
In the internal `getMostUsed` function, make the initial value of `result` more recognizeable as a failure value. Otherwise, when reading the function to try to understand it, it looks like `result` is usually expected to be a number. It is actually expected to be a string like `'s4'`.

This changes only an internal function. The external API remains the same.