Refactor index.js: extract functions, rename variables, etc.
I extracted functions so the logic of the exported function at the bottom is at a consistent, high level of abstraction. Some of the functions have doc comments that make their inputs and outputs clearer.

I renamed the variable `result` to make its type clearer and so as not to imply it’s the final result of the whole function. And I renamed `getMostUsed` to `getMostUsedKey` to make its return value clearer.

I made constants for the indent types 'space' and 'tab' and used them where possible instead of the less-readable values 's' and 't'. Now 's' and 't' are only used to simplify the encoding and decoding of map keys.